requirements_linux.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
appdirs=1.4.4=pypi_0
attrs=22.1.0=pypi_0
blas=1.0=mkl
bottleneck=1.3.5=py39h7deecbd_0
brotlipy=0.7.0=py39h27cfd23_1003
ca-certificates=2022.07.19=h06a4308_0
cattrs=22.2.0=pypi_0
certifi=2022.9.24=py39h06a4308_0
cffi=1.15.1=py39h74dc2b5_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
cryptography=37.0.1=py39h9ce1e76_0
exceptiongroup=1.0.0rc9=pypi_0
fftw=3.3.9=h27cfd23_1
greenlet=1.1.1=py39h295c915_0
idna=3.4=py39h06a4308_0
inflect=6.0.2=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
joblib=1.2.0=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
ncurses=6.3=h5eee18b_3
nose2=0.12.0=pypi_0
numexpr=2.8.3=py39h807cd23_0
numpy=1.23.3=py39h14f4228_0
numpy-base=1.23.3=py39h31eccc5_0
openssl=1.1.1q=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.4=py39h6a678d5_0
pip=22.2.2=py39h06a4308_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.10.2=pypi_0
pymongo=3.12.0=py39h295c915_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py39h06a4308_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.13=haa1d7c7_2
python-dateutil=2.8.2=pyhd3eb1b0_0
pytz=2022.1=py39h06a4308_0
pyyaml=6.0=pypi_0
readline=8.1.2=h7f8727e_1
requests=2.28.1=py39h06a4308_0
requests-cache=0.9.6=pypi_0
scipy=1.9.1=py39h14f4228_0
setuptools=63.4.1=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlalchemy=1.4.39=py39h5eee18b_0
sqlite=3.39.3=h5082296_0
statsbomb=0.3.0=pypi_0
tk=8.6.12=h1ccaba5_0
typing-extensions=4.4.0=pypi_0
tzdata=2022e=h04d1e81_0
url-normalize=1.4.3=pypi_0
urllib3=1.26.12=py39h06a4308_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.6=h5eee18b_0
zlib=1.2.13=h5eee18b_0
